[Refactor] scroll 이벤트 Refactoring (2023/09/14)
