+++ src/hooks/index.js
@@ -1,5 +1,4 @@
 export { default as useChannel } from './useChannel';
 export { default as useInfiniteFeed } from './useInfiniteFeed';
-export { default as useObserve } from './useObserve';
-export { default as useScroll } from './useScroll';
+export { default as useObserveScroll } from './useObserveScroll';
 export { default as useTitle } from './useTitle';
